Update Readme.md
Optional update for the documentation for systems where the `tools.go` doesn't link the binaries as expected. 

While the installation of the mentioned tools do go through, the binaries aren't linked as expected (at least on a M1 Air 2020), so proposing this addition 
for cases where someone might need to install them manually.